pyproject.toml
[project]
name = "swiss_chess"
version = "0.1.0"
description = "This application is a Swiss chess tournament manager."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "pandas>=1.3.5,<=1.5.4",
    "numpy==1.26.4",
    "matplotlib>=3.5.0",
    "scipy>=1.7.0",
    "scikit-image>=0.25.1",
    "streamlit>=1.37.0",
    "python-chess>=1.999",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.13.0",
    "pytest>=7.1.2",
    "pytest-cov>=3.0.0",
    "pre-commit==2.21.0",
    "mypy==1.6.0",
    "black==22.3.0",
    "flake8==4.0.1",
    "ruff==0.1.6"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
